Minor changes. - Rename EntityTypes.DROPPED_ITEM to ITEM - Clarify some BlockSnapshot javadocs - Use Vector3d for position in EntitySnapshotBuilder - Centralize DataQueries used in Location as public static final

+++ src/main/java/org/spongepowered/api/entity/EntitySnapshotBuilder.java
@@ -24,6 +24,7 @@
  */
 package org.spongepowered.api.entity;
 
+import com.flowpowered.math.vector.Vector3d;
 import com.flowpowered.math.vector.Vector3i;
 import org.spongepowered.api.data.ImmutableDataBuilder;
 import org.spongepowered.api.world.storage.WorldProperties;
@@ -62,7 +63,7 @@ public interface EntitySnapshotBuilder extends ImmutableDataBuilder<EntitySnapsh
      * @param position The Vector3i representing the coordinates
      * @return This builder, for chaining
      */
-    EntitySnapshotBuilder position(Vector3i position);
+    EntitySnapshotBuilder position(Vector3d position);
 
     /**
      * Copies over data from an {@link Entity}.
